refactor(core): move implementation classes to infra package
Relocate core implementation classes (OpenAiAnswerProvider, RobotScreenCapturer, and GlobalHotkeyBinder) to the infra package for improved separation of concerns and better project structure. Updated imports accordingly in related files.

No functional changes introduced.
